+++ src/components/ac-map/ac-map.component.ts
@@ -9,7 +9,7 @@ import { LabelDrawerService } from '../../services/drawers/label-drawer/label-dr
 import { PolylineDrawerService } from '../../services/drawers/polyline-drawer/polyline-drawer.service';
 import { PointDrawerService } from '../../services/drawers/point-drawer/point-drawer.service';
 import { ArcDrawerService } from '../../services/drawers/arc-drawer/arc-drawer.service';
-import { ViewersManagerService } from '../../services/viewers-service/viewers-manager.service';
+import { MapsManagerService } from '../../services/maps-manager/maps-manager.service';
 import { EllipseDrawerService } from '../../services/drawers/ellipse-drawer/ellipse-drawer.service';
 import { PolygonDrawerService } from '../../services/drawers/polygon-drawer/polygon-drawer.service';
 import { KeyboardControlService } from '../../services/keyboard-control/keyboard-control.service';
@@ -21,7 +21,7 @@ import { SceneMode } from '../../models/scene-mode.enum';
  * Every layer should be tag inside ac-map tag
  *
  * Accessing cesium viewer:
- * 1. acMapComponent.getViewer()
+ * 1. acMapComponent.getMap()
  * 2. Use ViewerManagerService.getCesiumViewer(mapId).
  *    mapId auto-generated string: 'default-map-id-[index]'
  *
@@ -53,37 +53,12 @@ import { SceneMode } from '../../models/scene-mode.enum';
   ]
 })
 export class AcMapComponent implements OnChanges, OnInit {
-  private static readonly DEFAULT_MINIMUM_ZOOM = 1.0;
-  private static readonly DEFAULT_MAXIMUM_ZOOM = Number.POSITIVE_INFINITY;
-  private static readonly DEFAULT_TILT_ENABLE = true;
-  private static defaultIdCounter = 1;
 
   /**
    * Disable default plonter context menu
    */
   @Input()
   disableDefaultPlonter = false;
-
-  // /**
-  //  * in meters
-  //  * @type {number}
-  //  */
-  // @Input()
-  // minimumZoom: number = AcMapComponent.DEFAULT_MINIMUM_ZOOM;
-  //
-  // /**
-  //  * in meters
-  //  * @type {number}
-  //  */
-  // @Input()
-  // maximumZoom: number = AcMapComponent.DEFAULT_MAXIMUM_ZOOM;
-  //
-  // /**
-  //  * Sets the enableTilt of screenSpaceCameraController
-  //  * @type {boolean}
-  //  */
-  // @Input()
-  // enableTilt: boolean = AcMapComponent.DEFAULT_TILT_ENABLE;
 
   /**
    * Set the id name of the map
@@ -111,7 +86,7 @@ export class AcMapComponent implements OnChanges, OnInit {
               private _cameraService: CameraService,
               private _elemRef: ElementRef,
               @Inject(DOCUMENT) private document: any,
-              private viewersManager: ViewersManagerService,
+              private mapsManagerService: MapsManagerService,
               private billboardDrawerService: BillboardDrawerService,
               private labelDrawerService: LabelDrawerService,
               private ellipseDrawerService: EllipseDrawerService,
@@ -119,7 +94,7 @@ export class AcMapComponent implements OnChanges, OnInit {
               private polygonDrawerService: PolygonDrawerService,
               private arcDrawerService: ArcDrawerService,
               private pointDrawerService: PointDrawerService,
-              private mapEventManager: MapEventsManagerService,
+              private mapEventsManager: MapEventsManagerService,
               private keyboardControlService: KeyboardControlService) {
     this.mapContainer = this.document.createElement('div');
     this.mapContainer.className = 'map-container';
@@ -129,11 +104,8 @@ export class AcMapComponent implements OnChanges, OnInit {
   }
 
   ngOnInit() {
-    // this._cesiumService.setMinimumZoom(this.minimumZoom);
-    // this._cesiumService.setMaximumZoom(this.maximumZoom);
-    // this._cesiumService.setEnableTilt(this.enableTilt);
-    this.viewersManager.setViewer(this.id, this.getCesiumViewer());
-    this.mapEventManager.init();
+    this.mapsManagerService.registerMap(this.id, this);
+    this.mapEventsManager.init();
     this.billboardDrawerService.init();
     this.labelDrawerService.init();
     this.ellipseDrawerService.init();
@@ -145,26 +117,33 @@ export class AcMapComponent implements OnChanges, OnInit {
   }
 
   ngOnChanges(changes: SimpleChanges): void {
-    if (changes['flyTo']) {
-      this._cameraService.flyTo(changes['flyTo'].currentValue);
-    }
     if (changes['sceneMode']) {
       this._cameraService.setSceneMode(changes['sceneMode'].currentValue);
+    }
+    if (changes['flyTo']) {
+      this._cameraService.cameraFlyTo(changes['flyTo'].currentValue);
     }
   }
 
   /**
-   * @returns {Viewer} map cesium viewer
+   * @returns {CesiumService} ac-map's cesium service
+   */
+  getCesiumSerivce() {
+    return this._cesiumService;
+  }
+
+  /**
+   * @returns {Viewer} map's cesium viewer
    */
   getCesiumViewer() {
     return this._cesiumService.getViewer();
   }
 
   /**
-   * @returns {Camera} map's scene's camera
+   * @returns {CameraService} map's scene's camera
    */
-  getCamera() {
-    return this._cameraService.getCamera();
+  getCameraService(): CameraService {
+    return this._cameraService;
   }
 
   /**
@@ -177,8 +156,8 @@ export class AcMapComponent implements OnChanges, OnInit {
   /**
    * @returns {MapEventsManagerService}
    */
-  getMapEventManager() {
-    return this.mapEventManager;
+  getMapEventsManager(): MapEventsManagerService {
+    return this.mapEventsManager;
   }
 
   /**