+++ src/services/maps-manager/maps-manager.service.ts
@@ -1,0 +1,30 @@
+import { Injectable } from '@angular/core';
+import { AcMapComponent } from '../../components/ac-map/ac-map.component';
+
+/**
+ *  The service manages `ac-map` instances. `ac-map` register itself to this service.
+ *  This allows retrieval of maps provided services outside of `ac-map` scope.
+ */
+@Injectable()
+export class MapsManagerService {
+  
+  private defaultIdCounter = 0;
+  private _Maps = new Map<string, any>();
+  
+  constructor() {
+  }
+  
+  getMap(id: string): AcMapComponent {
+    return this._Maps.get(id);
+  }
+  
+  registerMap(id: string, acMap: AcMapComponent) {
+    const mapId = id ? id : this.generateDefaultId();
+    this._Maps.set(mapId, acMap);
+  }
+  
+  private generateDefaultId(): string {
+    this.defaultIdCounter++;
+    return 'default-map-id-' + this.defaultIdCounter;
+  }
+}